Added new function
Function void upcust( ) . has added
for calculating bird flight distance between every customer and uploader center
